+++ stacking/base.py
